add ethValidationFailed
